backend/pyproject.toml
[project]
name = "backend"
version = "1.0.0"
description = ""
requires-python = ">=3.14"
dependencies = [
    "authx>=1.4.3",
    "environs>=14.5.0",
    "fastapi[standard]>=0.123.5",
    "jwt>=1.4.0",
    "pydantic-settings>=2.12.0",
    "pyjwt[crypto]>=2.10.1",
    "stp-database",
    "urllib3>=2.5.0",
]

[dependency-groups]
dev = [
    "pytest>=7.4.3",
    "mypy>=1.8.0",
    "ruff>=0.2.2",
    "pre-commit>=3.6.2",
    "types-passlib>=1.7.7.20240106",
    "coverage>=7.4.3",
]

[tool.uv.sources]
stp-database = { git = "https://github.com/STP-Team/stp-database" }

[tool.ruff.lint.pydocstyle]
convention = "google"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.mypy]
strict = true
exclude = ["venv", ".venv", "alembic"]
ignore_missing_imports = true

[tool.ruff]
target-version = "py310"
exclude = ["alembic"]

[tool.ruff.lint]
select = [
    "E", # pycodestyle errors
    "W", # pycodestyle warnings
    "F", # pyflakes
    "I", # isort
    "B", # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
    "ARG001", # unused arguments in functions
    "T201", # print statements are not allowed
]
ignore = [
    "E501", # line too long, handled by black
    "B008", # do not perform function calls in argument defaults
    "W191", # indentation contains tabs
    "B904", # Allow raising exceptions without from e, for HTTPException
]

[tool.ruff.lint.pyupgrade]
# Preserve types, even if a file imports `from __future__ import annotations`.
keep-runtime-typing = true

[tool.coverage.run]
source = ["app"]
dynamic_context = "test_function"

[tool.coverage.report]
show_missing = true
sort = "-Cover"

[tool.coverage.html]
show_contexts = true

[tool.hatch.build.targets.wheel]
packages = ["app"]
exclude = ["README.md", ".github/"]
